requirements.txt
streamlit
openai>=1.0
psycopg2-binary
pytesseract
Pillow

fastapi
uvicorn
